Product,Product Category and Product Picture Implemented

+++ DiscountManagement.Application.Contracts/CustomerDiscount/DefineCustomerDiscount.cs
@@ -6,7 +6,7 @@ namespace DiscountManagement.Application.Contracts.CustomerDiscount
     public class DefineCustomerDiscount
     {
         public long ProductId { get; set; }
-        public double DiscountRate { get; set; }
+        public int DiscountRate { get; set; }
         public string Reason { get; set; }
         public string StartDate { get; set; }
         public string EndDate { get; set; }
