+++ DiscountManagement.Domain/CustomerDiscountAgg/CustomerDiscount.cs
@@ -7,14 +7,14 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
     {
 
         public long ProductId { get; private set; }
-        public double DiscountRate { get; private set; }
+        public int DiscountRate { get; private set; }
         public string Reason { get; private set; }
         public DateTime StartDate { get; private set; }
         public DateTime EndDate { get; private set; }
 
     
 
-        public CustomerDiscount(long productId, double discountRate,
+        public CustomerDiscount(long productId, int discountRate,
             string reason, DateTime startDate, DateTime endDate)
         {
             ProductId = productId;
@@ -23,7 +23,7 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
             StartDate = startDate;
             EndDate = endDate;
         }
-        public void Edit(long productId, double discountRate,
+        public void Edit(long productId, int discountRate,
             string reason, DateTime startDate,DateTime endDate)
         {
             ProductId = productId;
